Add support for logging to a log file in the qt standard state directory
By default, logging to file is disabled. It can be enabled via the
configuration menu. Support for filtering by log level severity and
choosing a different path to save the  log files has also been added.

The log file implements a simple roller that saves only the two most
recent sets of logs generated by flameshot. This should stop log
messages from eating users' disk.

We use cmake to check the version of QT6. If greater than or equal to
6.7, we set the default log directory to the standard App State
directory. If not, we use the app local data directory